Added check for array and has cards

+++ js/settings.js
@@ -268,32 +268,55 @@ $("#admin-remove-column").click(function() {
     //note keeping column order as an extensible feature
     //columndata = JSON of columns
     var rest = [];
+    var colid;
     for(i=0; i<columndata.length; i++){
-        if(columndata[i].columnname != columnname){
-            rest.push(columndata[i]);
+        if(columndata[i].columnname == columnname){
+            colid = columndata[i].columnID;
         }
     }
-    for(k=0; k<rest.length; k++){
-        if(rest[k].columnID != k+1){
-            rest[k].columnID = k+1;
-        } 
-        if(rest[k].columnorder != k+1){
-            rest[k].columnorder = k+1;
+    
+    var hascolumns = false;
+    for(i=0; i<carddata.length; i++){
+        if(carddata[i]["info"].columnID == colid){
+            hascolumns = true;
         }
     }
-    //rest is what we save
-    localStorage.removeItem("columns");
-    var colstring = JSON.stringify(rest);
-    localStorage.setItem("columns", colstring);
-    //now push new data to the server
-    var upl = new XMLHttpRequest;
-    upl.open("POST", "./asc.php", true);
-    upl.setRequestHeader("Content-Type", "application/json");
-    upl.send(colstring);
-    //done
-    columndata = rest;
-    alert("Removed column: " + columnname);
-    location.reload();
-//    location.reload();
+    if(hascolumns){
+        var yesno;
+        vex.dialog.confirm({
+            message: 'This column has some cards, are you sure you want to delete?',
+            callback: function(value) {
+                yesno = value;
+            }
+        });
+    }
+    if(yesno){
+        for(i=0; i<columndata.length; i++){
+            if(columndata[i].columnname != columnname){
+                rest.push(columndata[i]);
+            }
+        }
+        for(k=0; k<rest.length; k++){
+            if(rest[k].columnID != k+1){
+                rest[k].columnID = k+1;
+            } 
+            if(rest[k].columnorder != k+1){
+                rest[k].columnorder = k+1;
+            }
+        }
+        //rest is what we save
+        localStorage.removeItem("columns");
+        var colstring = JSON.stringify(rest);
+        localStorage.setItem("columns", colstring);
+        //now push new data to the server
+        var upl = new XMLHttpRequest;
+        upl.open("POST", "./asc.php", true);
+        upl.setRequestHeader("Content-Type", "application/json");
+        upl.send(colstring);
+        //done
+        columndata = rest;
+        alert("Removed column: " + columnname);
+        location.reload();
+    }
 });
